Very basic benchmark. Refs #167.
One of the main reasons for Yampa to have changed very little, and for
so many PRs to take so long to integrate, is trying to make sure that we
do not seriously impact performance or break the API when we do that.

This commit implements a very, very basic benchmark that is integrated
in the cabal file.